+++ packages/engine/Source/Shaders/Materials/WaterMaskMaterial.glsl
@@ -1,0 +1,15 @@
+uniform vec4 waterColor;
+uniform vec4 landColor;
+
+czm_material czm_getMaterial(czm_materialInput materialInput)
+{
+    czm_material material = czm_getDefaultMaterial(materialInput);
+
+    vec4 outColor = mix(landColor, waterColor, materialInput.waterMask);
+    outColor = czm_gammaCorrect(outColor);
+
+    material.diffuse = outColor.rgb;
+    material.alpha = outColor.a;
+
+    return material;
+}